[untracked] initial commit of add and accept friend
