+++ api4/api.go
@@ -253,6 +253,7 @@ func Init(configservice configservice.ConfigService, globalOptionsFunc app.AppOp
 	api.InitTermsOfService()
 	api.InitGroup()
 	api.InitAction()
+	api.InitFriend()
 
 	root.Handle("/api/v4/{anything:.*}", http.HandlerFunc(api.Handle404))
